Add dotenv and axios as dependencies
In order to avoid hard coding access keys, dotenv is used to parse
environment variable declared in a file (.env) and populate Node's
global variable process.env. Placing .env in the same directory as the
Dockerfile and docker-compose.yml also makes it access its contents.
An example file has been added (.env.example).
Axios is added as the HTTP client for the project.