Add router and finish base funcionality
A router.js file was created to contain the endpoint logic (previosly
in app.js) wrapped in the Express Router to begin a better separation
of concerns. This file may be removed or divided in the future.
The API logic is basically done, but isn't structured as self-contained
components/modules yet. It's important to mention that requests to the
Giphy API are triggered recipe by recipe.